Break getDictionary into mutliple files and move them to src/service-worker/dictionaries
