+++ packages/scrabble-solver/src/service-worker/dictionaries/constants.ts
@@ -1,0 +1,3 @@
+export const DICTIONARY_CACHE = 'dictionary-api-cache';
+
+export const DICTIONARY_CACHE_MAX_AGE = 30 * 24 * 60 * 60 * 1000;
