+++ packages/scrabble-solver/src/service-worker/dictionaries/index.ts
@@ -1,0 +1,1 @@
+export { default as getDictionary } from './getDictionary';
